+++ nichtparasoup/imagecrawler/__init__.py
@@ -30,12 +30,14 @@ class KnownImageCrawlers(object):
         from .picsum import Picsum
         from .reddit import Reddit
         from .instagram import InstagramHashtag, InstagramProfile
+        from .pr0gramm import Pr0gramm
         return dict(
             Echo=Echo,
             Picsum=Picsum,
             Reddit=Reddit,
             InstagramProfile=InstagramProfile,
             InstagramHashtag=InstagramHashtag,
+            Pr0gramm=Pr0gramm,
         )
 
     def __init__(self, entries: Iterable[EntryPoint]) -> None:  # pragma: no cover